Bot_Categorias/requirements.txt
anyio==4.7.0
APScheduler==3.11.0
asyncio==3.4.3
certifi==2024.12.14
charset-normalizer==3.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
mysql-connector-python==9.2.0
pyaes==1.6.1
pyasn1==0.6.1
pyTelegramBotAPI==4.25.0
python-telegram-bot==21.9
requests==2.32.3
rsa==4.9
schedule==1.2.2
sniffio==1.3.1
SQLAlchemy==2.0.36
Telethon==1.38.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
validators==0.34.0
